Port the entire software to KConfigXT instead of our handmade config
This cleans a lot of files and we do not need to depend on a
barely-tested library anymore. While I do not like to write things in
xml, depending on KConfigXT allow us some nice features such as multiple
config files for profiles, and gracefully import of older configs.

It also allow us to stop requiring the preferences.h/.cpp auto-generated
files in our buildsystem, because the windows/mac bug on the CMake
function is no more.

Signed-off-by: Tomaz Canabrava <tcanabrava@kde.org>