Fix semantic rules for system libraries
The issue was that the previous implementation would match any file
that contains a file that ends with the same name as a given system
file. For instance BDE's open source library has a file named
'bdlpcre_regex.h', which matches with 'regex.h'.

It can still match local project files that fully matches the system
file names, but then again, this could potentially be an issue in the
project that has such files, while including them with <> instead of "".

Signed-off-by: Tarcisio Fischer <tarcisio.fischer@codethink.co.uk>