Do not construct uneeded std::filesystem::paths
std::filesystem::weakly_canonical expensive,
and we were constructing it multiple times for the same path.
now we do this only once on the constructor of the classes
that would use it.

This further reduces the time spend on a physical parse
more ~5 seconds on a single-thread while analyzing codevis