Remove "temporary" ifdef
Issue is that while running multiples instances of Codevis, it may fail
randomly if we don't ensure the database schema will be available. If we
ever need to remove this for some reason (e.g. updating the database),
then we need to recognize that the schema changed before updating it.

Signed-off-by: Tarcisio Fischer <tarcisio.fischer@codethink.co.uk>