Changes the logical parser to take in consideration the filename to disambiguate free function names
Free function's qualified name are now persisted as
"<funcName>@<filename>". This is not an ideal solution, as there
may be two files with same name with functions of same name defined in
different places, but it is a good heuristic for most projects.

Future implementations may need to take in consideration the compilation
and/or linkage context to resolve and disambiguate function calls in a
more formal way.

Signed-off-by: Tarcisio Fischer <tarcisio.fischer@codethink.co.uk>