Fix finding compiler header files on macos
The original code appended a single path, but newer versions
of the compilers in mac started to behave differently, including
an <iostream> would make codevis go hairwire.

This approach is the same one as on linux, removing a bunch
of ifdefs and making the code easier to read.

A new unittest was added to make sure this won't get
broken again.