Fix CPP Tool on Windows
This was painful.
Windows has a different way to handle files,
and we have three different libraries that opens
and handles them. Qt, LLVM, and the Standard library
Each one of those libraries prefer a specific way
to open files on windows and returns us paths in
different formats.

Making sure we store the correct format is like
chasing an hiperactive kid through the playground